LayoutTests/accessibility/gtk/media-controls-panel-title-expected.txt


This tests that the media controls panel has the correct name and role

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mainBlock.role is 'AXRole: AXGroup'
PASS mainBlock.title is 'AXTitle: '
PASS element.role is 'AXRole: AXVideo'
PASS element.title is 'AXTitle: '
PASS element.role is 'AXRole: AXToolbar'
PASS element.title is 'AXTitle: Video Playback'
PASS element.role is 'AXRole: AXAudio'
PASS element.title is 'AXTitle: '
PASS element.role is 'AXRole: AXToolbar'
PASS element.title is 'AXTitle: Audio Playback'
PASS successfullyParsed is true

TEST COMPLETE
